Add files via upload
My code for the TMLE estimator.  A few notes:
a) We need to provide a justification for the algorithms used in the library. I included them because these were the ones we worked in class with.
b) I wasn't able to get the 'ltmle' package to work, it kept throwing an error. I tried to fix it, but the couple of solutions I had didn't seem to work. We do have a TMLE estimate, so I gave up last night at 3AM.